+++ src/refactoring/hooks/index.ts
@@ -2,4 +2,5 @@ export * from "./useAddCoupon.ts";
 export * from "./useAddProduct.ts";
 export * from "./useCart.ts";
 export * from "./useCoupon.ts";
+export * from "./useEditProduct.ts";
 export * from "./useProduct.ts";
